Started implementing an option to save sequences for later use. -Created container model -Created mapping in the controller -Added jQuery to send a POST to controller
TODO: Stylize the sequence (perhaps add a bootstrap table with a *move*,
and *delete* buttons on each entry). Result textarea also disappears
when an entry is saved, values in each text area is gone (lost request
models). Be able to reset the session. Be able to download text file of
saved entry (with option of including the reverse complement as well as
the original in the file). Add visualization between the original and
reverse complement sequence. Add feature to receive DNA through BLAST
search and analyze the reverse complement. UI enhancements